+++ src/utils/timeUtil.h
@@ -5,8 +5,8 @@ class TimeUtil {
 public:
     static float deltatime;
     static float lastFrameTime;
-    static float frameRate;
-    static float frameRate1hz;
+    static int frameRate;
+    static int frameRate1hz;
     static float TimeAccumulator;
     static float time;
     static float priorTime; // for time based effects with motion vectors
